chg: move debug-server from n64 to ares
