+++ ares/ares/ares.cpp
@@ -1,5 +1,6 @@
 #include <ares/ares.hpp>
 #include <ares/debug/debug.cpp>
+#include <ares/debug-server/server.cpp>
 #include <ares/node/node.cpp>
 #include <ares/resource/resource.cpp>
