+++ ares/n64/GNUmakefile
@@ -1,7 +1,6 @@
 ares.components += sm5k
 
 ares.objects += ares-n64-memory
-ares.objects += ares-n64-debug-server
 ares.objects += ares-n64-system
 ares.objects += ares-n64-cartridge
 ares.objects += ares-n64-cic
@@ -19,26 +18,25 @@ ares.objects += ares-n64-cpu
 ares.objects += ares-n64-rsp
 ares.objects += ares-n64-rdp
 
-$(object.path)/ares-n64-memory.o:       $(ares.path)/n64/memory/memory.cpp
-$(object.path)/ares-n64-system.o:       $(ares.path)/n64/system/system.cpp
-$(object.path)/ares-n64-debug-server.o: $(ares.path)/n64/system/debug-server.cpp
-$(object.path)/ares-n64-cartridge.o:    $(ares.path)/n64/cartridge/cartridge.cpp
-$(object.path)/ares-n64-cic.o:          $(ares.path)/n64/cic/cic.cpp
-$(object.path)/ares-n64-controller.o:   $(ares.path)/n64/controller/controller.cpp
-$(object.path)/ares-n64-dd.o:           $(ares.path)/n64/dd/dd.cpp
-$(object.path)/ares-n64-sp.o:           $(ares.path)/n64/sp/sp.cpp
-$(object.path)/ares-n64-dp.o:           $(ares.path)/n64/dp/dp.cpp
-$(object.path)/ares-n64-mi.o:           $(ares.path)/n64/mi/mi.cpp
-$(object.path)/ares-n64-vi.o:           $(ares.path)/n64/vi/vi.cpp
-$(object.path)/ares-n64-ai.o:           $(ares.path)/n64/ai/ai.cpp
-$(object.path)/ares-n64-pi.o:           $(ares.path)/n64/pi/pi.cpp
-$(object.path)/ares-n64-pif.o:          $(ares.path)/n64/pif/pif.cpp
-$(object.path)/ares-n64-ri.o:           $(ares.path)/n64/ri/ri.cpp
-$(object.path)/ares-n64-si.o:           $(ares.path)/n64/si/si.cpp
-$(object.path)/ares-n64-rdram.o:        $(ares.path)/n64/rdram/rdram.cpp
-$(object.path)/ares-n64-cpu.o:          $(ares.path)/n64/cpu/cpu.cpp
-$(object.path)/ares-n64-rsp.o:          $(ares.path)/n64/rsp/rsp.cpp
-$(object.path)/ares-n64-rdp.o:          $(ares.path)/n64/rdp/rdp.cpp
+$(object.path)/ares-n64-memory.o:     $(ares.path)/n64/memory/memory.cpp
+$(object.path)/ares-n64-system.o:     $(ares.path)/n64/system/system.cpp
+$(object.path)/ares-n64-cartridge.o:  $(ares.path)/n64/cartridge/cartridge.cpp
+$(object.path)/ares-n64-cic.o:        $(ares.path)/n64/cic/cic.cpp
+$(object.path)/ares-n64-controller.o: $(ares.path)/n64/controller/controller.cpp
+$(object.path)/ares-n64-dd.o:         $(ares.path)/n64/dd/dd.cpp
+$(object.path)/ares-n64-sp.o:         $(ares.path)/n64/sp/sp.cpp
+$(object.path)/ares-n64-dp.o:         $(ares.path)/n64/dp/dp.cpp
+$(object.path)/ares-n64-mi.o:         $(ares.path)/n64/mi/mi.cpp
+$(object.path)/ares-n64-vi.o:         $(ares.path)/n64/vi/vi.cpp
+$(object.path)/ares-n64-ai.o:         $(ares.path)/n64/ai/ai.cpp
+$(object.path)/ares-n64-pi.o:         $(ares.path)/n64/pi/pi.cpp
+$(object.path)/ares-n64-pif.o:        $(ares.path)/n64/pif/pif.cpp
+$(object.path)/ares-n64-ri.o:         $(ares.path)/n64/ri/ri.cpp
+$(object.path)/ares-n64-si.o:         $(ares.path)/n64/si/si.cpp
+$(object.path)/ares-n64-rdram.o:      $(ares.path)/n64/rdram/rdram.cpp
+$(object.path)/ares-n64-cpu.o:        $(ares.path)/n64/cpu/cpu.cpp
+$(object.path)/ares-n64-rsp.o:        $(ares.path)/n64/rsp/rsp.cpp
+$(object.path)/ares-n64-rdp.o:        $(ares.path)/n64/rdp/rdp.cpp
 
 ifeq ($(vulkan),true)
   ifeq ($(filter shortest-stem,${.FEATURES}),)
