+++ desktop-ui/desktop-ui.hpp
@@ -5,6 +5,7 @@
 using namespace hiro;
 
 #include <ares/ares.hpp>
+#include <ares/debug-server/server.hpp>
 #include <mia/mia.hpp>
 
 #include <nall/instance.hpp>
